+++ packages/ai-jsx/src/core/node.ts
@@ -12,6 +12,7 @@ import { Logger } from './log.js';
 /** A context that is used to render an AI.JSX component. */
 export interface ComponentContext extends RenderContext {
   logger: Logger;
+  isAppendOnlyRender: boolean;
 }
 
 /** Represents a single AI.JSX component. */
@@ -32,7 +33,7 @@ export interface Element<P> {
   /** The component properties. */
   props: P;
   /** A function that renders this {@link Element} to a {@link Renderable}. */
-  render: (renderContext: RenderContext, logger: Logger) => Renderable;
+  render: (renderContext: RenderContext, logger: Logger, isAppendOnlyRender: boolean) => Renderable;
   /** The {@link RenderContext} associated with this {@link Element}. */
   [attachedContextSymbol]?: RenderContext;
 }
@@ -76,17 +77,30 @@ export function makeIndirectNode<T extends object>(value: T, node: Node): T & In
 
 /** @hidden */
 export function withContext(renderable: Renderable, context: RenderContext): Element<any> {
-  function SwitchContext() {
-    return renderable;
+  if (isElement(renderable)) {
+    if (renderable[attachedContextSymbol]) {
+      // It's already been bound to a context; don't replace it.
+      return renderable;
+    }
+
+    const elementWithContext = {
+      ...renderable,
+      [attachedContextSymbol]: context,
+    };
+    Object.freeze(elementWithContext);
+    return elementWithContext;
   }
 
-  const elementWithContext = {
-    ...(isElement(renderable) ? renderable : createElement(SwitchContext, null)),
-    [attachedContextSymbol]: context,
-  };
-
-  Object.freeze(elementWithContext);
-  return elementWithContext;
+  // Wrap it in an element and bind to that.
+  return withContext(
+    createElement(
+      function SwitchContext({ children }) {
+        return children;
+      },
+      { children: renderable }
+    ),
+    context
+  );
 }
 
 /** @hidden */
@@ -119,7 +133,7 @@ export function createElement<P extends { children: C | C[] }, C>(
   const result = {
     tag,
     props: propsToPass,
-    render: (ctx, logger) => tag(propsToPass, { ...ctx, logger }),
+    render: (ctx, logger, isAppendOnlyRender) => tag(propsToPass, { ...ctx, logger, isAppendOnlyRender }),
   } as Element<P>;
   Object.freeze(propsToPass);
   Object.freeze(result);